Extract devcontainer export package
