+++ cli/python/base_devcontainer/__init__.py
@@ -1,0 +1,3 @@
+"""Dev Containers export support for Base manifests."""
+
+from __future__ import annotations
